Reorganize cosmos-client-components modules
