+++ crates/cosmos/cosmos-client-components/src/components/ibc_client.rs
@@ -4,8 +4,8 @@ use hermes_relayer_components::chain::traits::components::connection_handshake_m
 use hermes_relayer_components::chain::traits::components::consensus_state_querier::ConsensusStateQuerierComponent;
 use hermes_relayer_components::chain::traits::components::create_client_message_builder::CreateClientMessageBuilderComponent;
 
-use crate::impls::connection_handshake_message::BuildCosmosConnectionHandshakeMessage;
-use crate::impls::create_client_message::BuildCosmosCreateClientMessage;
+use crate::impls::client::create_client_message::BuildCosmosCreateClientMessage;
+use crate::impls::connection::connection_handshake_message::BuildCosmosConnectionHandshakeMessage;
 use crate::impls::query_client_state::QueryCosmosClientStateFromChainHandle;
 use crate::impls::query_consensus_state::QueryCosmosConsensusStateFromChainHandle;
