+++ crates/cosmos/cosmos-client-components/src/impls/mod.rs
@@ -1,13 +1,8 @@
 pub mod ack_packet_message;
 pub mod ack_packet_payload;
-pub mod channel_handshake_message;
-pub mod channel_handshake_payload;
-pub mod connection_handshake_message;
-pub mod connection_handshake_payload;
-pub mod create_client_message;
-pub mod create_client_payload;
-pub mod init_channel_options;
-pub mod init_connection_options;
+pub mod channel;
+pub mod client;
+pub mod connection;
 pub mod packet_fields;
 pub mod packet_from_ack;
 pub mod query_block;
@@ -28,5 +23,3 @@ pub mod send_messages_as_tx;
 pub mod timeout_packet_message;
 pub mod timeout_packet_payload;
 pub mod types;
-pub mod update_client_message;
-pub mod update_client_payload;
