+++ crates/solomachine/solomachine-relayer/src/impls/chain/component.rs
@@ -1,7 +1,7 @@
 use cgp_core::prelude::*;
 use hermes_cosmos_client_components::impls::packet_fields::CosmosPacketFieldReader;
 use hermes_cosmos_client_components::impls::types::chain::ProvideCosmosChainTypes;
-use hermes_cosmos_client_components::impls::update_client_message::BuildCosmosUpdateClientMessage;
+use hermes_cosmos_client_components::impls::client::update_client_message::BuildCosmosUpdateClientMessage;
 use hermes_relayer_components::chain::traits::components::channel_handshake_message_builder::ChannelHandshakeMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::components::channel_handshake_payload_builder::ChannelHandshakePayloadBuilderComponent;
 use hermes_relayer_components::chain::traits::components::client_state_querier::ClientStateQuerierComponent;
