+++ crates/sovereign/sovereign-client-components/src/sovereign/impls/mod.rs
@@ -1,2 +1,3 @@
 pub mod client;
+pub mod connection;
 pub mod types;
